Deliverable_2

+++ Backend/demo/demo/src/main/java/com/example/demo/Repository/UserBudgetRepository.java
@@ -3,5 +3,5 @@ package com.example.demo.Repository;
 import com.example.demo.Model.UserBudget;
 import org.springframework.data.jpa.repository.JpaRepository;
 
-public interface UserBudgetRepository extends JpaRepository<UserBudget, Integer> {
+public interface UserBudgetRepository extends JpaRepository<UserBudget, Long> {
 }
